refactor: move tied-weight helpers to shared
