+++ pyproject.toml
@@ -475,9 +475,6 @@ modules = [
 ignore_imports = [
     # Dion needs the FSDP DP mesh resolver so its optimizer mesh matches parameter sharding.
     "nemo_automodel.components.optim.dion -> nemo_automodel.components.distributed.mesh_utils",
-    # TP restores the embedding/lm_head alias before FSDP records parameter
-    # ownership; the tied-embedding helpers are owned by the checkpoint component.
-    "nemo_automodel.components.distributed.parallelizer -> nemo_automodel.components.checkpoint.utils",
     "nemo_automodel.components.distributed.optimized_tp_plans -> nemo_automodel.components.models.llama.model",
     "nemo_automodel.components.distributed.optimized_tp_plans -> nemo_automodel.components.models.gemma4_moe.model",
     "nemo_automodel.components.distributed.optimized_tp_plans -> nemo_automodel.components.models.mistral3_vlm.model",
