+++ tests/unit_tests/utils/test_checkpoint_utils.py
@@ -18,6 +18,13 @@ import pytest
 import torch.nn as nn
 
 import nemo_automodel.components.checkpoint.utils as checkpoint_utils
+from nemo_automodel.shared import tied_weights
+
+
+def test_checkpoint_utils_reexports_shared_tied_weight_helpers():
+    """Checkpoint callers retain their existing imports while shared code owns the helpers."""
+    assert checkpoint_utils.ensure_tied_lm_head is tied_weights.ensure_tied_lm_head
+    assert checkpoint_utils.has_local_tied_lm_head is tied_weights.has_local_tied_lm_head
 
 
 def test_is_tied_word_embeddings_prefers_top_level_value():
